Add explicit instantiations for IVF-PQ search kernels used in tests (#2212)
Compilation of IVF-PQ search kernels can be time consuming. In `libraft.so` the compilation is done in parallel for kernels without filtering and with `int64_t` index type.

We have test with `uint32_t` index type as well as tests for `bitset_filter` with both 32 and 64 bit index types. This PR adds explicit template instantiations for the test. This way we avoid repeated compilation of the kernels with filter and this also enables parallel compilation of the `compute_similarity` kernel for different template types. The kernels with these additional type parameters are not added to `libraft.so`, only linked together with the test executable. 

Note that this PR does not increase the number of compiled kernels, but it enables to compile them in parallel.

Authors:
  - Tamas Bela Feher (https://github.com/tfeher)

Approvers:
  - Artem M. Chirkin (https://github.com/achirkin)
  - Ben Frederickson (https://github.com/benfred)

URL: https://github.com/rapidsai/raft/pull/2212

+++ cpp/src/neighbors/detail/ivf_pq_compute_similarity_float_fp8_true_filt32.cu
@@ -25,4 +25,4 @@ instantiate_raft_neighbors_ivf_pq_detail_compute_similarity_select(
   float,
   raft::neighbors::ivf_pq::detail::fp_8bit<5u COMMA true>,
   raft::neighbors::filtering::ivf_to_sample_filter<
-    int64_t COMMA raft::neighbors::filtering::none_ivf_sample_filter>);
+    uint32_t COMMA raft::neighbors::filtering::none_ivf_sample_filter>);
